+++ Libplanet.Explorer/ExplorerStartup.cs
@@ -1,8 +1,5 @@
 #nullable disable
 using GraphQL;
-using GraphQL.Server;
-using GraphQL.SystemTextJson;
-using GraphQL.Types;
 using Libplanet.Action;
 using Libplanet.Explorer.GraphTypes;
 using Libplanet.Explorer.Interfaces;
@@ -55,9 +52,7 @@ namespace Libplanet.Explorer
             services.TryAddSingleton<ExplorerQuery<T>>();
             services.TryAddSingleton<LibplanetExplorerSchema<T>>();
 
-            services.AddGraphQL()
-                    .AddSystemTextJson()
-                    .AddGraphTypes(typeof(LibplanetExplorerSchema<T>));
+            services.AddGraphQL(builder => builder.AddLibplanetExplorerSchema<T>());
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)