+++ Libplanet.Explorer/GraphTypes/ActionType.cs
@@ -1,13 +1,9 @@
 #nullable disable
 using System;
-using System.Buffers;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Text;
 using System.Text.Json;
 using Bencodex;
-using Bencodex.Types;
 using GraphQL;
 using GraphQL.Types;
 using Libplanet.Action;
@@ -19,16 +15,10 @@ namespace Libplanet.Explorer.GraphTypes
     {
         public ActionType()
         {
-            Field<NonNullGraphType<StringGraphType>>(
-                name: "Raw",
-                description: "Raw Action data ('hex' or 'base64' encoding available.)",
-                arguments: new QueryArguments(
-                    new QueryArgument<StringGraphType>
-                    {
-                        DefaultValue = "hex",
-                        Name = "encode",
-                    }),
-                resolve: ctx =>
+            Field<NonNullGraphType<StringGraphType>>("Raw")
+                .Description("Raw Action data ('hex' or 'base64' encoding available.)")
+                .Argument<string>("encode", false, arg => arg.DefaultValue = "hex")
+                .Resolve(ctx =>
                 {
                     var codec = new Codec();
                     var encoded = codec.Encode(ctx.Source.PlainValue);
@@ -48,27 +38,22 @@ namespace Libplanet.Explorer.GraphTypes
                                 "It supports only 'hex' or 'base64'.";
                             throw new ExecutionError(msg);
                     }
-                }
-            );
+                });
 
-            Field<NonNullGraphType<StringGraphType>>(
-                name: "Inspection",
-                description: "A readable representation for debugging.",
-                resolve: ctx => ctx.Source.PlainValue.Inspect(loadAll: true)
-            );
+            Field<NonNullGraphType<StringGraphType>>("Inspection")
+                .Description("A readable representation for debugging.")
+                .Resolve(ctx => ctx.Source.PlainValue.Inspect(loadAll: true));
 
-            Field<NonNullGraphType<StringGraphType>>(
-                name: "json",
-                description: "A JSON representaion of action data",
-                resolve: ctx =>
+            Field<NonNullGraphType<StringGraphType>>("json")
+                .Description("A JSON representation of action data")
+                .Resolve(ctx =>
                 {
                     var converter = new Bencodex.Json.BencodexJsonConverter();
                     var buffer = new MemoryStream();
                     var writer = new Utf8JsonWriter(buffer);
                     converter.Write(writer, ctx.Source.PlainValue, new JsonSerializerOptions());
                     return Encoding.UTF8.GetString(buffer.ToArray());
-                }
-            );
+                });
 
             Name = "Action";
         }