+++ Libplanet.Explorer/GraphTypes/ByteStringType.cs
@@ -1,7 +1,6 @@
-#nullable disable
 using System;
-using GraphQL.Language.AST;
 using GraphQL.Types;
+using GraphQLParser.AST;
 
 namespace Libplanet.Explorer.GraphTypes
 {
@@ -12,7 +11,7 @@ namespace Libplanet.Explorer.GraphTypes
             Name = "ByteString";
         }
 
-        public override object Serialize(object value)
+        public override object? Serialize(object? value)
         {
             return value switch
             {
@@ -22,7 +21,7 @@ namespace Libplanet.Explorer.GraphTypes
             };
         }
 
-        public override object ParseValue(object value)
+        public override object? ParseValue(object? value)
         {
             switch (value)
             {
@@ -35,13 +34,7 @@ namespace Libplanet.Explorer.GraphTypes
             }
         }
 
-        public override object ParseLiteral(IValue value)
-        {
-            return value switch
-            {
-                StringValue stringValue => ParseValue(stringValue.Value),
-                _ => null,
-            };
-        }
+        public override object? ParseLiteral(GraphQLValue? value) =>
+            value is GraphQLStringValue v ? ParseValue((string)v.Value) : null;
     }
 }